refactor: HintItem들을 hint.item package 내부로 분리
Hint와 HintItem 구분을 위함
