+++ src/main/java/baseball/domain/hint/item/HintItem.java
@@ -1,4 +1,4 @@
-package baseball.domain.hint;
+package baseball.domain.hint.item;
 
 public abstract class HintItem {
