+++ src/main/java/baseball/domain/hint/item/Nothing.java
@@ -1,4 +1,4 @@
-package baseball.domain.hint;
+package baseball.domain.hint.item;
 
 public class Nothing extends HintItem {
