+++ src/test/java/baseball/service/hint/item/BallServiceImplTest.java
@@ -1,61 +1,63 @@
-package baseball.service.hint;
+package baseball.service.hint.item;
 
 import baseball.domain.NumberBaseball;
-import baseball.domain.hint.HintItem;
-import baseball.domain.hint.Strike;
+import baseball.domain.hint.item.Ball;
+import baseball.domain.hint.item.HintItem;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class StrikeServiceImplTest {
+public class BallServiceImplTest {
 
-    private final StrikeServiceImpl strikeHintServiceImpl = StrikeServiceImpl.getInstance();
+    private final BallServiceImpl ballServiceImpl = BallServiceImpl.getInstance();
 
     @Test
     void create() {
         NumberBaseball computerBaseball = NumberBaseball.createBaseball("123");
         NumberBaseball inputBaseball = NumberBaseball.createBaseball("456");
 
-        HintItem strikeHint = strikeHintServiceImpl.create(computerBaseball, inputBaseball);
+        HintItem ballHint = ballServiceImpl.create(computerBaseball, inputBaseball);
 
-        assertTrue(strikeHint instanceof Strike);
+        assertTrue(ballHint instanceof Ball);
     }
 
+
     @Test
-    void count_zeroStrike() {
+    void count_zeroBall() {
         //given
         NumberBaseball computerBaseball = NumberBaseball.createBaseball("123");
         NumberBaseball inputBaseball = NumberBaseball.createBaseball("456");
 
         //when
-        int count = strikeHintServiceImpl.count(computerBaseball, inputBaseball);
+        int count = ballServiceImpl.count(computerBaseball, inputBaseball);
 
         //then
         assertThat(count).isZero();
     }
 
     @Test
-    void count_zeroStrike_differentPlace() {
-        //given
-        NumberBaseball computerBaseball = NumberBaseball.createBaseball("123");
-        NumberBaseball inputBaseball = NumberBaseball.createBaseball("451");
-
-        //when
-        int count = strikeHintServiceImpl.count(computerBaseball, inputBaseball);
-
-        //then
-        assertThat(count).isZero();
-    }
-
-    @Test
-    void count_oneStrike() {
+    void count_zeroBall_samePlace() {
         //given
         NumberBaseball computerBaseball = NumberBaseball.createBaseball("123");
         NumberBaseball inputBaseball = NumberBaseball.createBaseball("156");
 
         //when
-        int count = strikeHintServiceImpl.count(computerBaseball, inputBaseball);
+        int count = ballServiceImpl.count(computerBaseball, inputBaseball);
+
+        //then
+        assertThat(count).isZero();
+    }
+
+    @Test
+    void count_oneBall() {
+        //given
+        NumberBaseball computerBaseball = NumberBaseball.createBaseball("123");
+        NumberBaseball inputBaseball = NumberBaseball.createBaseball("451");
+
+        //when
+        int count = ballServiceImpl.count(computerBaseball, inputBaseball);
 
         //then
         assertThat(count).isOne();
